Moved fixture to one spot
